L157 Physics & Tank movement.

+++ BattleTank/Source/BattleTank/Public/Tank.h
@@ -55,5 +55,5 @@ private:
 
 
 
-	double LastFireTime = 0;
+	float LastFireTime = 0;
 };
